requirements.txt
openai
tweepy
apscheduler
load_dotenv